fix: move the destinationInputToConfigurer to conversions
